pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aurelia"
version = "0.1.0"
description = "Automated discovery from the command line"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "google-genai>=1.0",
    "mcp>=1.0",
    "structlog",
    "gitpython",
    "anyio",
]

[project.scripts]
aurelia = "aurelia.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/aurelia"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
